Create rangefinder2.py
This collection of routines causes the After Sight Model 1 to monitor the GPIO of the raspberry pi. The GPIO look for the rotary encoder button press, and start a timer to toggle the vibration/range finder. Other GPIO look for low battery, power switch position, and the plug/unplug or external power. the serial rx/tx of the raspberry pi are used to take input from the maxbotix rangefinder module.