added blinders,zoom,and foveal mapping to raspivoice settings. Added triple vibration on startup, removed --speak from raspivoice execution
added config items for blinders zoom and foveal mapping, added those
choices to the menus.
the triple vibration on startup is to let users know the device is
working even if their audio is not
removed --speak from raspivoice execution: this change allows the
keyboard to be used without crashing raspivoice. However you have to
double tap +,-, or enter to get the desired result.