Added Facial Detection
I have tried to make a way to classify position of the face in a 3x3
grid It doesn't work perfectly, but I found it reliable for detecting
when a face is present, and if I am centered on the face with the
camera. Thinking about adding a depth feature as the rectangle size
changes with distance on the average adult head.