Implemented Automatic Core Update
Added Update command to settings menu. it does a git pull, and then runs
all core update scripts to rebuild/recopy raspivoice, teradeep, and the
facialdetection. The distancesensor is updated as part of the home
directory scripts as it doesn't need it's own folder.
installdeps is meant to keep the devices up to date. Once an apt-get
package is installed, a hold is placed on it to prevent the loss or
unwanted upgrade in the future. If you do want to upgrade a package,
then make a one time script related to the update number to make that
happen.
changed rangefinder frequencies to between 7-9khz. People complained
that 14khz was way too high to hear for older people.
update scripts had commands to force installation added so as not to
hang the system on questions requiring yes/no answers